+++ scripts/publishPackages.js
@@ -15,7 +15,7 @@ function checkForNewVersion(cwd) {
   const markdown = fs.readFileSync(cwd + '/CHANGELOG.md', { encoding: 'UTF-8' })
   const nextVersion = markdown.split('## ')[1].split('\n')[0].trim()
 
-  if (/^\d\.\d\.\d(-.+)?$/.test(nextVersion) && version !== nextVersion) {
+  if (/^\d+\.\d+\.\d+(-.+)?$/.test(nextVersion) && version !== nextVersion) {
     return nextVersion
   }
 }
@@ -32,6 +32,7 @@ for (const path of packages) {
     const nextVersion = checkForNewVersion(path)
     if (nextVersion) {
       const context = { cwd: path, stdio: 'inherit' }
+      tags.push(`${getPackageName(path)}@${nextVersion}`)
       cp.execSync('npm version ' + nextVersion, context)
       cp.execSync(
         'npm set //registry.npmjs.org/:_authToken ' +
@@ -39,7 +40,6 @@ for (const path of packages) {
         { cwd: path }
       )
       cp.execSync('npm publish --access=public', context)
-      tags.push(`${getPackageName(path)}@${nextVersion}`)
     }
   } catch (e) {
     console.error(e)
@@ -47,4 +47,4 @@ for (const path of packages) {
   }
 }
 
-process.stdout.write(tags.join(' '))
+process.stdout.write(tags.map((tag) => tag.trim()).join(' '))